secao19_banco_dados_python/requirements.txt
autopep8==1.5.3
mysql-connector==2.2.9
pycodestyle==2.6.0
toml==0.10.1
